requirements.txt
amqp==5.1.1
anyjson==0.3.3
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
Brotli==1.0.9
celery==5.2.7
celery-progress==0.1.3
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
cssselect2==0.7.0
Django==4.1.4
django-celery-beat==2.4.0
django-celery-results==2.4.0
django-chartjs==2.3.0
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-phonenumber-field==7.0.2
django-timezone-field==5.0
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.2.1
et-xmlfile==1.1.0
ez-setup==0.9
fonttools==4.38.0
html5lib==1.1
idna==3.4
kombu==5.2.4
linkedin-api==1.1.0
numpy==1.24.0
openpyxl==3.0.10
pandas==1.5.2
phonenumbers==8.13.4
Pillow==9.3.0
prompt-toolkit==3.0.37
psycopg2-binary==2.9.5
pycparser==2.21
pydyf==0.5.0
pymongo==3.12.1
pyphen==0.13.2
python-crontab==2.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7
redis==4.5.1
requests==2.28.1
six==1.16.0
sqlparse==0.2.4
tinycss2==1.2.1
typing==3.7.4.3
tzdata==2022.7
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.6
weasyprint==58.0
webencodings==0.5.1
zopfli==0.2.2
opencage==2.1.0